Move ml.utils.output to utils and rename it to classifier_utils
